PROJ-48_ForgotRecoveryPasswordApi refactor: move validation rules in model folder and reduce import paths
