+++ features/newPassword/api/index.ts
@@ -1,2 +1,1 @@
 export * from './api'
-export * from './ui'
